projection: out casted to input values dtype

+++ art/attacks/evasion/projected_gradient_descent/projected_gradient_descent_pytorch.py
@@ -504,6 +504,6 @@ class ProjectedGradientDescentPyTorch(ProjectedGradientDescentCommon):
             else:  # Non-convex optim
                 raise NotImplementedError('Values of `norm_p < 1` are currently not supported with `suboptimal=False`')
 
-        values = values_tmp.reshape(values.shape)
+        values = values_tmp.reshape(values.shape).to(values.dtype)
 
         return values
